Polishing

+++ File-O-Matic/FileRepresentationWindow.cs
@@ -14,6 +14,7 @@ namespace File_O_Matic
     public partial class FileRepresentationWindow : Form
     {
         private String filename;
+        private Boolean saved = false;
 
         public FileRepresentationWindow(String filename)
         {
@@ -22,7 +23,14 @@ namespace File_O_Matic
             this.filename = filename;
             this.filenameLabel.Text = filename;
 
-            
+            try
+            {
+                this.pictureBox.Image = Image.FromFile(filename);
+            } catch (Exception)
+            {
+
+                this.pictureBox.Image = Image.FromFile("File.jpeg");
+            }
         }
 
 
@@ -41,15 +49,40 @@ namespace File_O_Matic
 
         private void CheckSave(object sender, FormClosingEventArgs e)
         {
+            if (!this.saved)
+            {
+                this.DoSave();
+            }
+        }
+
+        private void DoSave()
+        {
             if (MessageBox.Show("Do you want to save this file?", "Save File", MessageBoxButtons.YesNo) == DialogResult.Yes)
             {
                 var savedialog = new SaveFileDialog();
-                
+
                 if (savedialog.ShowDialog() == DialogResult.OK)
                 {
-                    System.IO.File.Move(this.filename, savedialog.FileName);
+                    if (!System.IO.File.Exists(this.filename))
+                    {
+                        if (MessageBox.Show("A file already exists with this name. Would you like to override it?", "Override File", MessageBoxButtons.YesNo) == DialogResult.Yes)
+                        {
+                            System.IO.File.Move(this.filename, savedialog.FileName);
+                            this.saved = true;
+                        }
+                    }
                 }
             }
+        }
+
+        private void FileRepresentationWindow_Load(object sender, EventArgs e)
+        {
+
+        }
+
+        private void button3_Click(object sender, EventArgs e)
+        {
+            this.DoSave();
         }
     }
 }
